New change on the function of revertedWith
Hi to the best Blockchain Community in the internet! Im currently doing the testing of withdraw function, and with the version of patrick, today (17/7/2022) the expect function will return this error: 
```
AssertionError: Expected transaction to be reverted with reason 'FundMe__NotOwner', but it reverted with a custom error
```

So apparently, chai made some changes and now we have to use revertedWithCustomError()! 
The first parameter should be the contract interface where the custom error is triggered and the second parameter should be the name of but in string type.

Thanks for this amazing course @PatrickAlphaC !!!